Move result to separate file
